The great refactor - moving all under Linter namespace
