+++ lib/linter/task.rb
