+++ lib/linter/version.rb
@@ -1,0 +1,3 @@
+module Linter
+  VERSION = "0.0.1"
+end
