--- lib/mas-lint.rb
@@ -1,1 +1,0 @@
-require 'mas/lint'
