--- lib/mas/lint/version.rb
@@ -1,5 +1,0 @@
-module Mas
-  module Lint
-    VERSION = "0.0.1"
-  end
-end
